Fix leaking task resources when nodes are deleted
When a node is deleted, its tasks are asked to restart, which involves
putting them into a desired state of Shutdown. However, the Allocator
will not deallocate a task which is not in an actual state of a terminal
state. Once a node is deleted, the only opportunity for its tasks to
recieve updates and be moved to a terminal state is when the function
moving those tasks to TaskStateOrphaned is called, 24 hours after the
node enters the Down state. However, if a leadership change occurs, then
that function will never be called, and the tasks will never be moved to
a terminal state, leaking resources.

With this change, upon node deletion, all of its tasks will be moved to
TaskStateOrphaned, allowing those tasks' resources to be cleaned up.

Additionally, as part of this backport, avoid using the gogo
types.TimestampNow function, which does not exist in the vendored
version.

Signed-off-by: Drew Erny <drew.erny@docker.com>
(cherry picked from commit 8467e6a43f9cee66f46efb457b2c22f46ae5da0b)
Signed-off-by: Sebastiaan van Stijn <github@gone.nl>