Moving swap options into `ResourceRequirements` instead of `ContainerSpec`s
It has been decided when discussing https://github.com/moby/moby/pull/37872
that swap options should progressively be moved to existing structs
meant to limit resource usage, and that the new field for swap should hold
the swap size itself, instead of the swap size + the memory.

This patch is a pre-req to unblock https://github.com/moby/moby/pull/37872
in the light of the new design.

Removing the 2 fields from `ContainerSpec` proto messages shouldn't be an
issue as these were never used in practice.

Signed-off-by: Jean Rouge <rougej+github@gmail.com>