Refactor CreateExtensionRequest
While working on the Resources code, I realized this API for creation,
where we pass only the allowed arguments instead of taking a whole
object and erroring if disallowed fields are present, is better.

This change breaks the protocol buffers, but it is acceptable because
there was never a user of the "old" protobufs.

Signed-off-by: Drew Erny <drew.erny@docker.com>