Use Service Placement Constraints in Enforcer
This patch updates the ConstraintEnforcer to use the placement
constraints from the current service spec rather than the placement
constraints from the current task spec because they may be outdated
in some cases (e.g., when the service was previously updated to
modify placement constrainst but the node to which the task was
scheduled still satisfies the constraints. If the node is updated
in a way which causes it to no longer satisfy the new constraints
then the task should be removed even if it still would satisfy the
original task constraints).

Docker-DCO-1.1-Signed-off-by: Josh Hawn <josh.hawn@docker.com> (github: jlhawn)