Only update non-terminal tasks on node removal.
When a node is removed, its tasks are set in state ORPHANED. This does
not need to be done for tasks that are already in a terminal state, and
if all tasks in all states are updated, the size of the transaction may
grow too large to process, and node removal becomes impossible.

This changes to only set non-terminal tasks to state ORPHANED, and
terminal tasks are left alone.

Cherry pick does not apply cleanly

Signed-off-by: Drew Erny <drew.erny@docker.com>
(cherry picked from commit d5df26594f5b74a9de41e376206c36abffc961fe)
Signed-off-by: Drew Erny <drew.erny@docker.com>