Refactor Jobs Orchestrators
It became evident in the process of writing the Global Jobs orchestrator
that the Orchestrators required by both Replicated and Global jobs are
essentially identical. This commit merges them into one combined
orchestrator, which dispatches to the appropriate Reconcilers to do the
actual work.

Unlike existing services, these orchestrators can be combined because
the requirements of jobs are much simpler than that of services.

Signed-off-by: Drew Erny <drew.erny@docker.com>