Support restarting failing jobs tasks
In order to make the jobs reconcilers work correctly with the restart
supervisor, they have been altered to never replace failed tasks
directly Replacing failed tasks is the purview of the restart
supervisor. The jobs reconcilers will only create new tasks when needed.

Additionally, this alters the behavior of the replicated job reconciler
with regards to slots -- each new task will get a new slot, and when the
job is completed, there will be a Completed task in each slot from 0 to
TotalCompletions-1.

Then, makes the tweaks necessary for the Restart Supervisor to support
Jobs, which are different from other services in that they deliberately
have a desired state of Completed.

Finally, wires up the replicated and global orchestrators to call the
restart supervisor to restart tasks that have failed.

Signed-off-by: Drew Erny <drew.erny@docker.com>