Verify network allocation on creating network
Current network allocation is run asynchronously with network (object) creation, when a user specifies an overlapped subnet, the network creation succeeds but this network is useless with empty IPAM config.

This commit makes network creation use a channel to wait for the result of network allocation from Allocator before return.
Because of this change, some _test.go files are modified.

Also added TestCreateNetworkOverlapIP to network_test.go.

Remove duplicated line in "make coverage" of direct.mk.

Signed-off-by: Xinfeng Liu <xinfeng.liu@gmail.com>