Remove t.Log from mock task controller
Previous, worker_test.go included a call to t.Log in the mock task
controller. This presented a problem: because goroutines spawned by a
test are not exited when the test completes, any holdover goroutines,
like the one spawned to close the task controller, live on. This creates
a problem when that goroutine calls t.Log on a test that has already
completed.

This t.Log call doesn't add enough information to bother worrying about.
Further, these goroutines, though they may outlive the test, don't
actually affect anything or matter, so there's no rush to fix the test
to ensure all goroutines exit.

Signed-off-by: Drew Erny <drew.erny@docker.com>