ca: remove uses of deprecated go-digest.Digest.Hex()
This function was in https://github.com/opencontainers/go-digest/commit/55f675811a1b915549933c64571fd86b2676ba76,
but the format of the GoDoc comments didn't follow the correct format, which
caused it not being picked up by tools as "deprecated".

This patch updates uses in the codebase to use the alternatives.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>