manager/state/raft/storage: MigrateSnapshot fix etcd v3.5.6 compatiblity
Commit https://github.com/etcd-io/etcd/commit/3644c9d67ba0316c1c8f7c7d96316bf739234381
changed the signature of fileutil.CreateDirAll to now require a go.uber.org/zap.Logger
to be passed.

Unfortunately that's a type (not an interface), which results in go.uber.org/zap.Logger
to be a direct dependency.

Looking at what fileutil.CreateDirAll does;

- It checks if each directory exists, and has the expected permissions (0700).
- It checks if each directory is empty.

Given that we already delete the directory (and terminate if that fails) before
we create the new location, that functionality is redundant, so we can just
as well use `os.MkdirAll()` to achieve the same.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>