Fix removing tasks when a jobs service is removed
Previously, the jobs orchestrators were missing the appropriate code to
handle the deletion of a service. As a result, when a service was
deleted, the tasks for that service would hang around indefinitely. It's
likely that on a leadership change or restart, the tasks would have been
deleted, but that's generally not an acceptable process as leadership
changes or restarts aren't guaranteed.

Fixes this issue by adding event handling for services, similar to
global or replicated services, that moves all tasks for a deleted
service into Remove state, flagging them for deletion.

Signed-off-by: Drew Erny <derny@mirantis.com>