m/a/cnmallocator: convert to new-style driver registration
The libnetwork driver registry and driver Init() functions were
deprecated in https://github.com/moby/moby/commit/28edc8e2d692cb47f1ec6bb32a6e346777e1250d
in favour of separate registries for network and IPAM drivers and
Register() functions for the drivers.

Refactor the network driver initializer list to be a map as the network
types should be unique, and it allows IsBuiltInDriver() to run in O(1)
time.

The remote drivers are a special case as they dynamically register
plugins with the driver registries when the plugins are initialized.
As they are the only drivers which need a PluginGetter, register the
remote drivers separately as a special case rather than trying to
continue shoehorning them into initializers and initIPAMDrivers().

There is no "remote" network type. However, Swarmkit has accepted it as
a valid network type ever since support for selectable network drivers
was first added in 5507383c809156ca7218f6140a328cde0aae2af1. This is a
bug, as attempting to create such a network would ultimately fail.
Special-casing the remote network driver incidentally fixes the bug as
it is no longer present in initializers, which is used as the source of
truth for IsBuiltinDriver().

Signed-off-by: Cory Snider <csnider@mirantis.com>