Add support for security opts on containers
Once, a long long time ago, we had aspirations of including a more
structured way to handle security permissions in Swarmkit.

Unfortunately, this never panned out. Rather than continuing to stagnate
on a might-be, this change simply add in support for the --security-opt
flag and its API equivalents directly. Perfect shouldn't be the enemy of
good enough.

Signed-off-by: Drew Erny <derny@mirantis.com>