manager: inject network allocator Provider
allocator.Allocator is ostensibly loosely coupled with the low-level
network allocator implementation by virtue of depending on an
abstraction: the networkallocator.NetworkAllocator interface. However,
in practice it also depends directly on the cnmallocator implementation.
The Allocator constructs a CNM network allocator directly, and code
in the allocator package directly calls functions in the cnmallocator
package. A similar issue exists within controlapi. While it does not
directly depend on cnmallocator, its validation logic contains deep
knowledge of what is and is not valid for the CNM network allocator to
consume.

Decouple manager.Manager and allocator.Allocator from the cnmallocator
implementation of the low-level network allocator by refactoring them to
depend on a new Provider interface instead of the cnmallocator
functions. Decouple controlapi.Server from the cnmallocator
implementation by refactoring it to depend on a new Validator interface
implemented by the low-level network allocators and lowering the
validation logic into the cnmallocator implementation. Construct the
cnmallocator Provider in the node.Node and inject it into the
manager.Manager.

Signed-off-by: Cory Snider <csnider@mirantis.com>