manager/allocator: move Config struct out of CNM
The network allocator configuration is all sourced from the Cluster
object and persisted in Raft. Any change to the configuration therefore
requires modifications to the Swarmkit API objects and manager logic.
Move the cnmallocator.NetworkConfig struct to networkallocator.Config to
reflect that Swarmkit, not the network allocator implementations,
dictates which configuration values are passed into the allocator.

Signed-off-by: Cory Snider <csnider@mirantis.com>