manager/allocator: delete cnmallocator package
It is only needed to integrate Swarmkit into Moby networking and no
Swarmkit package imports it. It is also the only package remaining which
imports from github.com/docker/docker. Delete it to break the module
dependency cycle with Moby. The code will be moved over to the
https://github.com/moby/moby repo to live closer to the libnetwork logic
that it is so tightly integrated into.

Signed-off-by: Cory Snider <csnider@mirantis.com>