swarmd: remove /v2/ from module path
The module path for swarmd is not valid for the location of the swarmd
module within the swarmkit repo. The go command resolves package paths
to the swarmkit/v2 module and fails when the packages don't exist.

    $ go run github.com/moby/swarmkit/v2/swarmd/cmd/swarm-rafttool@master
    go: github.com/moby/swarmkit/v2/swarmd/cmd/swarm-rafttool@master: module github.com/moby/swarmkit/v2@master found (v2.0.0-20240227173239-911c97650f2e), but does not contain package github.com/moby/swarmkit/v2/swarmd/cmd/swarm-rafttool

(This is only a problem when swarmd is not the main module.)

Fix the swarmd module's path to align with the module source location so
that `go run` can once again be used to invoke swarm-rafttool and the
other commands.

Signed-off-by: Cory Snider <csnider@mirantis.com>
Signed-off-by: plaurent <patrick@saint-laurent.us>