manager/controlapi: implement New constructor with functional arguments
Commit bc6b233149a2f1dbf75adebaa742254abc4e2ef7 introduced a new feature
to provide control API handlers for GetNetwork and ListNetworks. As part
of this change, the NewServer constructor's signature was extended with
a new argument.

Adding a new argument is a breaking change, and would require a major
version update of the module (although it looks like there's no external
consumers, other than through vendored files).

This patch implements a new constructor that uses functional argument,
which allows expanding the controlapi server with new functionality
in future without breaking the API.

The current implementation in this patch moves all options to functional
arguments, but some arguments are required, and may be better set through
positional arguments. For example, the codebase performs no nilness checks
for the memory store or securityconfig. Other options, such as the raft
node should likely be required as well, but the codebase does handle those
more gracefully (the server would check for a non-nil store, with one
exception, which was added as part of this patch).

Driver provider already assumed for a pluginProvider to be missing, but
did not yet account for no provider to be present at all; for this, the
code was adjusted both through a nil check, and by setting an empty driver
provider.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>