api: afford passing params to OnGetNetwork hook
Provide a facility for the application to pass options into its
OnGetNetwork hook. Add an Any-typed field to the GetNetwork and
ListNetworks Control API requests that gets passed into the
NetworkAllocator's callback function.

Pass the callback the typeUrl and value as separate parameters instead
of the *Any so that the OnGetNetworker interface does not need to change
when we migrate away from gogo protobuf.

Signed-off-by: Cory Snider <csnider@mirantis.com>